feat: 🎸 say-as tag
